Issue-292: Refactored circularrevealfragment.
