+++ circularrevealfragment/src/main/java/com/kirchhoff/circularrevealfragment/fragments/BaseCircularRevealFragment.kt
@@ -2,9 +2,10 @@ package com.kirchhoff.circularrevealfragment.fragments
 
 import android.os.Bundle
 import android.view.View
+import android.view.ViewAnimationUtils
+import android.view.animation.DecelerateInterpolator
 import com.example.core.ui.BaseFragment
-import com.kirchhoff.circularrevealfragment.utils.ExitWithAnimation
-import com.kirchhoff.circularrevealfragment.utils.startCircularReveal
+import kotlin.math.hypot
 
 internal abstract class BaseCircularRevealFragment : BaseFragment(), ExitWithAnimation {
 
@@ -16,5 +17,35 @@ internal abstract class BaseCircularRevealFragment : BaseFragment(), ExitWithAni
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         view.startCircularReveal(circularRevealFromLeft())
+    }
+
+    private fun View.startCircularReveal(fromLeft: Boolean) {
+        addOnLayoutChangeListener(object : View.OnLayoutChangeListener {
+            override fun onLayoutChange(
+                v: View,
+                left: Int,
+                top: Int,
+                right: Int,
+                bottom: Int,
+                oldLeft: Int,
+                oldTop: Int,
+                oldRight: Int,
+                oldBottom: Int
+            ) {
+                v.removeOnLayoutChangeListener(this)
+                val cx = if (fromLeft) v.left else v.right
+                val cy = v.bottom
+                val radius = hypot(right.toDouble(), bottom.toDouble()).toFloat()
+                ViewAnimationUtils.createCircularReveal(v, cx, cy, 0f, radius).apply {
+                    interpolator = DecelerateInterpolator(2f)
+                    duration = ANIMATION_DURATION
+                    start()
+                }
+            }
+        })
+    }
+
+    companion object {
+        private const val ANIMATION_DURATION = 1000L
     }
 }
